Renamed python module to cuckoo_time_translator_python
This is now identical to its package/project name and apparently just the way it was expected to be.

Addressing second part of #49 (2.)
